fix(unittests): make xlat table tests run for LPA2 and non LPA2 cases
This patch refactors all the xlat lib unittests so they can run
twice, once with FEAT_LPA2 supported and once with that feature
unsupported.

This patch also fixes a small bug on the xlat library in which
the minimum level for a block descriptor was wrongly calculated
when creating translation tables.

+++ plat/host/host_test/include/test_helpers.h
@@ -102,7 +102,7 @@ int test_helpers_unregister_cb(enum cb_ids id);
 /*
  * Return a random value within [min, max] range.
  */
-int test_helpers_get_rand_in_range(int min, int max);
+int test_helpers_get_rand_in_range(long min, long max);
 
 /*
  * Defines whether we expect or not an assert to fail.
